backend/requirements.txt
Django==5.2.10
whitenoise[brotli]==6.8.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==24.3
Pillow==11.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
channels==4.0.0
channels-redis==4.2.0
celery==5.3.4
redis==5.0.1
daphne==4.1.0